Added googletest
Roughly based on this:
https://stackoverflow.com/questions/35998856/how-can-i-use-google-test-with-my-project-that-builds-via-autotools/36000856#36000856

This integrates googletest/gtest with the regular autotools "make check"
system. googletest then puts nicer output in test-suite.log than we
would normally get there.

This is not particularly useful in this project because so many of the
tests are actually compile-time tests with static_assert().